fastapi_project/requirements.txt
altair==5.4.0
annotated-types==0.7.0
anyio==3.7.1
async-generator==1.10
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
dnspython==2.6.1
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.112.1
fastapi-login==1.10.1
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
motor==3.5.1
narwhals==1.4.2
numpy==2.0.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pillow==10.4.0
plotly==5.23.0
protobuf==5.27.3
pyarrow==17.0.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.38.2
streamlit==1.37.1
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
trio==0.21.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
